Putting things in the right place.
